Melhorias e correção de bugs
[#106165912 #103549422]Correção de links, implementação de novos botões
"cancelar" nos formulários, melhorias visuais e atendendo a solicitação
da issue #61 foram corrigidas algumas regras de validação de formulários
e alteradas as mensagens de erro de validação.

+++ retaguarda/database/migrations/2015_10_28_100527_create_prices_table.php
@@ -15,10 +15,10 @@ class CreatePricesTable extends Migration
         Schema::create('prices', function (Blueprint $table) {
             $table->increments('id');
             $table -> integer('min_time');
-             $table -> integer('un_price');
+             $table -> decimal('un_price');
              $table -> integer('un_time');
             $table -> integer('discount_sellers');
-            $table -> integer('max_price');
+            $table -> decimal('max_price');
         });
     }
